Fixed scopes display
Previously, whenever a variable was assigned for the first time after not being assigned during its initialization, the variable would be passed to every scope. This was not intentional, as our only desire is for the variable to be updated where it actually exists. Therefore, some notable changes were made to a couple of features. Now the symbol table's ScopeIDs are of the type Int. Moreover, the bindST, enterScopeST and exitScopeST suffered some changes, especially bindST (this was quite brutal). This is a clear demonstration of how not making functions easily changeable can hinder (or even bottleneck) the development of the entire structure. This unwanted "feature" has been fixed, so we're one step closer to finishing the current stage of development for this Symbol Table.